requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
pyjwt
redis
python-multipart
email-validator